Move computed collections to kredits service
They should be available from any place with access to the service.
